requirements.txt
realesrgan==0.3.0
basicsr>=1.4.2
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
facexlib>=0.2.2
gfpgan>=1.3.0